# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VulnOps Hub
  description: Vulnerability-operations control plane — modular monolith
  version: 0.1.0
paths:
...
  /api/v1/organizations/{org_id}/assets/import:
    post:
      tags:
      - assets
      summary: Import Assets
      description: 'Import a CSV asset inventory (CMDB export).


        Header row required; recognized columns: hostname (required identity),

        name, criticality, environment, owner, internet_exposure, type,

        business_service_id, service, service_name.

        Existing assets are matched by hostname alias and updated in place;

        unknown hostnames create assets with a hostname alias; ambiguous alias

        collisions are skipped for review, never merged.'
      operationId: import_assets_api_v1_organizations__org_id__assets_import_post
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Import Assets Api V1 Organizations  Org Id  Assets
                  Import Post
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ProblemDetails:
      properties:
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          type: string
          title: Code
        detail:
          type: string
          title: Detail
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Fields
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      title: ProblemDetails
      description: Stable error envelope shared by authenticated API operations.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      description: OIDC access token
      scheme: bearer
      bearerFormat: JWT